Minor changes in main.c: now it's possible to select the heuristic function to be used for A Star. Sequential A Star function protoype has been changed such that the desired heuristic is passed as argument. All heuristic functions now return int; Haversine body has been changed by adding conversion from coordinates to longitude and latitude values. In sequential A Star, PQ init has modified by allocating an array of size G->V; as a consequence, the reallocation in PQinsert has been deactivated (it was giving too much errors).

+++ src/Heuristic.h
@@ -4,7 +4,7 @@
 #include "ST.h"
 
 int Hcoord(Coord coord1, Coord coord2);
-double Hhaver(Coord coord1, Coord coord2);
-double Hdijkstra(Coord coord1, Coord coord2);
+int Hhaver(Coord coord1, Coord coord2);
+int Hdijkstra(Coord coord1, Coord coord2);
 
 #endif
